+++ CHANGES.rst
@@ -5,5 +5,10 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Show the lead image of Note items in the "All content" (full view)
+  listing of collections and folders, and enable the ``plone.leadimage``
+  behaviour on the Note type so new sites get the lead image field.
+  [kaerumy]
+
 - Initial release.
   [kaerumy]